hv: trusty: save/restore tsc for sworld when suspend/resume
TSC would be reset to 0 when enter suspend state. This would
fail the secure timer checking in secure world because secure
world leverage the TSC as source of secure timer which should be
increased monotonously.

This patch save/restore in secure world suspend/resume path to
guarantee the mono increasing TSC of secure world.

Signed-off-by: Qi Yadong <yadong.qi@intel.com>